refactor: packages - moved all the radix-ui primitives from the dashboard to @repo/ui since they are related to shadcn components and will be reused by the web app - moved getLanguageName and getLanguageColor to @repo/ui/utils because those two utilities are related to the ui
