+++ apps/dashboard/package.json
@@ -12,11 +12,6 @@
   "dependencies": {
     "@fontsource-variable/inter": "^5.1.1",
     "@hookform/resolvers": "^5.0.1",
-    "@radix-ui/react-checkbox": "^1.3.2",
-    "@radix-ui/react-dropdown-menu": "^2.1.5",
-    "@radix-ui/react-label": "^2.1.4",
-    "@radix-ui/react-popover": "^1.1.6",
-    "@radix-ui/react-slot": "^1.2.0",
     "@repo/trpc": "*",
     "@repo/common": "*",
     "@repo/ui": "*",
@@ -25,12 +20,8 @@
     "@trpc/client": "^11.1.0",
     "@trpc/server": "^11.1.0",
     "@trpc/tanstack-react-query": "^11.1.0",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
     "d3-hierarchy": "^3.1.2",
-    "date-fns": "^4.1.0",
     "js-md5": "^0.8.3",
-    "lucide-react": "^0.474.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-error-boundary": "^5.0.0",
@@ -38,8 +29,6 @@
     "react-router": "^7.5.1",
     "react-spinners": "^0.17.0",
     "recharts": "^2.15.1",
-    "tailwind-merge": "^2.6.0",
-    "tailwindcss-animate": "^1.0.7",
     "zod": "^3.24.3",
     "zustand": "^5.0.3"
   },
