+++ packages/ui/src/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_COLOR = "HSL(334, 90%, 51%)";
