+++ packages/ui/src/utils/getLanguageColor.ts
@@ -1,5 +1,5 @@
-import { DEFAULT_COLOR } from "./constants";
-import languagesAttributes from "./colors.json";
+import { DEFAULT_COLOR } from "../constants";
+import languagesAttributes from "../colors.json";
 const getLanguageColor = (languageSlug: string) => {
   let languageColor: string = "";
   try {
